Working on a fix with those 429 pesky too many request
I found out that the IA may flag certain URLs being saved as a success without showing an error message if things like saving twitter (currently, resulting "Please try again in ~1 min. Crawling this host is paused because they notified us that are overloaded right now." frequently) if it errors out 429, AND will show https://web.archive.org/web/YYYYMMDDhhmmss/<siteURL> on column D instead of the error message. Therefore, I added a filter and a sort for HTTP status codes.

I also changed the format, InputList[Index][1] now contains a subarray, that way the array are organized and not in their own main "directory" (sorted so that [Index][0] is the google sheets, [Index][1] is the JS flagging)