We only need to push once, and not keep looking through the list
Although WBGSInfoList[Index][1][0] and WBGSInfoList[Index][1][1] NEVER contain multiple status and therefore not be pushed into the list multiple times resulting duplicates, it's the fact that once a match is found, it gets pushed into the list as expected, BUT when that inner loop finishes, it still searches and check if the other filters matches, and all of that following matches will fail, which we don't need to check anymore. Therefore I added a check that would terminates the loop to save time.