Learned about callback functions and filter
Now utilizing javascript's array filter feature: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter the arguments is like sort, but has one parameter. When you have a function name WITHOUT the parenthesis (like this: FunctionName not FunctionName(), that can be inserted as an argument without calling it). I would believe that you can put ANYTHING as a parameter for a function

I wanted to have a requirement set by the user that if archive status *OR* HTTP status have been met will be included in the list. This is not possible if you have two separate filter functions. Therefore, you can specify if you wanted OR or AND on both flags.