minor fix and feature added
if you have capturing groups in string.split, the separators themselves will be included in the array between elements that were split. This means every other item in the array are just line breaks in IA_WBGS_Checking.html rather than the desired data when entering the GS table causing RemoveInvalidData to remove N items where N is the number of URLs plus if there is a new line break at the end.

That was found when attempting to add a feature to display the line number position of each row in GS.