GreaseMonkey_GeneralLinksRecursiveWindow_WIP.js
This is the version of GreaseMonkey_GeneralLinks.js but is capable of extracting links in all the sub-windows in a frame/iframe. Almost finished, had one issue: if only the subwindow (not the main window) loads a different document, it will not detect a load-finish, the scrolling no longer works.

This is because the source of placing the event listeners is at the main window, meaning these will add event listeners based on the main window.