update
added failsafe measures if there are too many nested frames that could potentially cause the control to call the recursive function too many times and exceed the stack size.

Also removed a random code I placed there just for debugging.

I also discovered when using the interval, and the framed window goes to another page, the interval keeps old window data, which now becomes a dead object. So what happens is that the event listner added to those windows that got unloaded remained in the JS's setInterval's Input_Window.document.

When you visit a normal page without any window frames, and you go to another page, the event listeners resets along with most of the values on the GM script. But if a page with window frames and you open different links, in a way that affects only the window frames, does not reset the entire page and the GM script.