Finally working!
This greasemonkey script is now ready to use. Let me know if you find any bugs on that. This is a tool that when used alongside GreaseMonkey_GeneralLinksRecursiveWindow_WIP.js will be a great AFK link extraction tool as an alternative to https://github.com/com-puter-tips/Links-Extractor but can scrape URLs that are not from the HTML itself but also from the JS.