working on a new script - Greasemonkey_NitterAutoNext
WIP though, breaks on searches, this basically "autoclicks" the "Load more" button so you can AFK while the other scripts are extracting links. Be careful though, even with auto-refreshes this may not extract every tweet, despite nitter, the front end of twitter, not use twitter's API, which that encumbered by the 7-day not show some tweets and other on-twitter restrictions.