update
Listcleaner and WBGS checking - "reset" button now works properly, I've haven't touch this for a long time and with so many updates, it fails to properly reset all the changes in the HTML. Also make it so if the list is empty for listcleaner or if there is no valid entry for WBGS, also clears the output.

WBGS checking now actually calculates the total by simply adding all the "sure" and "unsure" states and not have an extra variable, this makes espically the loop for each URL entry not having to count on BOTH variables when a state is detected.

Listcleaner's unused functions are removed