Something I long forgotten - fragment identifiers
Added a setting to ignore fragment identifiers (https://en.wikipedia.org/wiki/URI_fragment ). When saving URLs containing a hash followed by any character before the end of string, if saved successfully, will display the URL without the fragment. Before this change, it will flag it as a redirect. Not sure if javascript-loaded pages can behave differently and load different content (which results in saving different data) which is why I made it an option than to make it always ignore fragments. No page so far I seen never have JS to conditionally load stuff on the HTML based on the fragment.