added another ignore-match - http vs https
One site long ago that I visited errors out if you try to view the https version of a site and the http version works fine. This is a reason why this is an option as well.

Also, changed how the code gets the page URL out of the WBM URL version, no longer using string search and substring (was been there since ages ago), now uses string replace.