WTF Elun musk
Saving tweets normally or through nitter no longer works because of a decision to lock tweets behind a login wall in a response to sites scraping content "aggressively": https://sea.mashable.com/tech/24669/twitter-now-blocks-visitors-from-viewing-tweets-and-profiles-unless-theyre-logged-in

On top of that, instead of 404ing on web crawlers, it will now display a message saying "Unverified web crawlers are not allowed".

This only applies to any twitter pages. The media URLs themselves remain working fine.