keyname added
Added another attribute to sort by if you happen to run into, say unsavable URLs but do not want to remove off the list, so that if you find an alternative URL to save and need to search. Also useful if you need to find URLs to extract links off of.