update
-Slight adjustments added to the checking tool. One of the missing feature that it had on the very old version long ago is now added in the new version was the green text saying all the URLs are saved. I also added an item count that I initially thought it is redundant is the number of items displayed, placed on top of the textarea output.

-The TwitterURLProcessor have some of the unchecked-by-default now be checked by default in case if saving pages become possible again, and added a button that unchecks all twitter pages. I'm not so sure if twitter is blocking not just twitter.com/* but also on media URLs under https://pbs.twimg.com/*

It's really a dark day for twitter, even with the new CEO replacing Elun Musk, Linda Yaccarino, is still suffering: https://www.techdirt.com/2023/10/06/extwitter-ad-revenue-continues-to-be-in-free-fall/

Some users are moving to project bluesky. Problem is:
"Currently, Bluesky is in beta, with registration only available to those with an invite, but the company has outlined its plans for eventual rollout to the general public"

Meaning like discord, it is invite-only, and not public (allow unregistered users; lurking, search engine indexing), thus WBM cannot save bluesky until "eventual rollout to the general public".