modernized WBGS checking tool
in beta right now. This more exploits the way that when writing data on each line rather than columns, because when you first split the spreadsheet, you MUST split it by lines .split(/(?:\n)(?=(?:[^"]|"[^"]*")*$)/) thus the index represents what line number, and THEN you split again by "\t" and that second index represents the column number. Because this is Row-major (each URL is one line), a foreach/map will work on this (since it takes the outmost index of the outer array if it were a 2D array, if it was that each COLUMN have a URL and the list is sideways, you would have to use for and increment the second index (and not use foreach/map) for each URL item rather than the first), or convert the sideway list into a vertical list so that the main index is each URL.