update
moving CheckAvailability to use objects as well as optimizations

also added a comment for developments (p tag using hidden attribute) letting them know that a setting that modifies how the checking system treats and flags stuff needs to count that as a change on the list. I renamed it to IsListOrModeChanged to clarify this.