update
haven't tested the start process pages yet since I'm still waiting on my 12+ hour long job-queued process.

Made sure that any URL substring no longer have the "h" at the beginning of the string because Firefox STILL truncates URLs and doesn't actually preserve them (if you copy the text or the log (via mouse-select-text), it will have the truncated version, while the link it goes to when clicked on uses the full URL, while on chrome, only the display is truncated, and copying the text, or the log, will copy the full URL)

I now added a timestamp info in the console log, to make tracking URLs at what time it has started (the time the URL appears at the start process page's bottom, not necessarily when the process transitions from an intially-queued state to traversing URLs).