yeah, doing reference node is the most STABLE way to obtain data
If you try to use the DOM tree from the root node, it will fail if there are hidden elements that would move the node containing all the posts around so it is a bad idea.

Next up is to extract posts while viewing a post and not being in a user profile page.