realize why that bug occurred...
If there is only a single post with no replies and is not a reply to anything above, the top code that tries to find a node containing all post will fail because there is no a href of the post. I also realized.

Still documenting the format of the dom tree post