"Reply to ..." crashes the script
If you scroll down fast enough, the post with "Reply to <User_title>" can sometimes still be in a loading state, as "Reply to ...", which causes function "GetPostContent" to error out at "NodeForDivs.OutputNode.getElementsByTagName" saying that it is not a function (NodeForDivs.OutputNode would be a TEXT NODE (pointing to the "Reply to") instead of an ELEMENT NODE). The "getElementsByTagName" only exists on element nodes, not text nodes.