thanks https://stackoverflow.com/a/12522758/11030779
So I was concerned about that furaffinity yesterday, and it got me worried, what if any JS method or function could get overriden and break any userscripts? Well, that is actually true, however, you can retrieve methods and functions via iframes, since it is a completely seperate document and its own window, that contains the original JS as if you loaded the document before any code would override the default methods and functions.