fixed a bug that causes post objects with missing data
So part of bsky update was half-loading posts that are offscreen. When this happens, certain elements are not on the HTML causing function DescendNode to return a failure.  So my idea is to just simply not allow half-loaded posts be saved in a list.

Another update is updating to the new layout of the profile content.