update
Month separator lines are now always drawn at the first rendered row on days whose days above it (7 days ago) is in the previous month. Before, if a month whose first day is not Sunday, and the week below it is the first to be rendered at the table's top, the days on that week whose day "above" (7 days ago, directly up on the calendar) it is the previous month will not have month separator lines drawn.