not sure why I deleted helperfunctions, and realized how important they are
these functions assume the argument "Node" has to be the LOWEST DOM tree node that includes just the quoted post (has to be a div, and 1 level lower will start getting elements containing the header, text and images). We can then adjust the caller of these functions's argument depending on the format.