For numbers that can represent fractions, better to use tofixed
"However, the toString() method doesn't directly use this most precise representation of the number value. Rather, the algorithm uses the least number of significant figures necessary to distinguish the output from adjacent number values."

Take for example, 1/3, realistically equals to 0.(3).

either not manuelly converting to a string or doing so using toString(10), would print out 0.3333333333333333. Assuming the integer part is not counted, that is 16 digits.

This is "the least number of significant figures necessary to distinguish the output from adjacent number values", which adjacent representable value is 0.333333333333333314829616256247390992939472198486328125, which starting at the 17th digit (immediately after the last "3", again not counting the integer digit) is when the digits are incorrect.

This is a problem that is solved using toFixed.