moar cleanup
The save statuses are now handled as a group and are centralized as an object initializer, this includes the color and the count. This makes should a new save status type is identified, simply edit the object "SaveStatusTypeList", the map function at "Main process, the result calculations at "Totals of the 3 group of types" and the HTML tables. I made the counter for each type more modular and seperate from the map function so that when you add a new type, you do not need to have <Counter>++ along with setting the "SaveStatus" (things that have a consistent pattern are in a seperate iteratable function).