Realized a mistake...
This method causes problems where if a not-fully-loaded post having a video is a reply chain in between 2 posts, ends up connecting the replies above and below it to each other, which we do not want. Instead Planning to include them in the PostGroup but not in the save. It will check in the save and when re-running that a non-fully-save will be replaced with a fully save but not the other way around